+++ Maui.Controls.UserDialogs/Shared/ConfirmConfig.cs
@@ -21,12 +21,20 @@ public class ConfirmConfig
 
     public static Color DefaultMessageColor { get; set; }
     public static Color DefaultTitleColor { get; set; }
-    public static double DefaultTitleFontSize { get; set; } = 20;
-    public static double DefaultMessageFontSize { get; set; } = 16;
+    public static float DefaultTitleFontSize { get; set; } = 20;
+    public static float DefaultMessageFontSize { get; set; } = 16;
+
+    /// <summary>
+    /// Android only
+    /// </summary>
     public static Color DefaultPositiveButtonTextColor { get; set; }
-    public static double DefaultPositiveButtonFontSize { get; set; } = 18;
+    public static float DefaultPositiveButtonFontSize { get; set; } = 18;
+
+    /// <summary>
+    /// Android only
+    /// </summary>
     public static Color DefaultNegativeButtonTextColor { get; set; }
-    public static double DefaultNegativeButtonFontSize { get; set; } = 18;
+    public static float DefaultNegativeButtonFontSize { get; set; } = 18;
 
     public string FontFamily { get; set; } = DefaultFontFamily;
     /// <summary>
@@ -52,12 +60,20 @@ public class ConfirmConfig
 
     public Color MessageColor { get; set; } = DefaultMessageColor;
     public Color TitleColor { get; set; } = DefaultTitleColor;
+
+    /// <summary>
+    /// Android only
+    /// </summary>
     public Color PositiveButtonTextColor { get; set; } = DefaultPositiveButtonTextColor;
+
+    /// <summary>
+    /// Android only
+    /// </summary>
     public Color NegativeButtonTextColor { get; set; } = DefaultNegativeButtonTextColor;
-    public double TitleFontSize { get; set; } = DefaultTitleFontSize;
-    public double MessageFontSize { get; set; } = DefaultMessageFontSize;
-    public double PositiveButtonFontSize { get; set; } = DefaultPositiveButtonFontSize;
-    public double NegativeButtonFontSize { get; set; } = DefaultNegativeButtonFontSize;
+    public float TitleFontSize { get; set; } = DefaultTitleFontSize;
+    public float MessageFontSize { get; set; } = DefaultMessageFontSize;
+    public float PositiveButtonFontSize { get; set; } = DefaultPositiveButtonFontSize;
+    public float NegativeButtonFontSize { get; set; } = DefaultNegativeButtonFontSize;
 
     public static bool DefaultUseYesNo { get; set; }
     public static string DefaultYes { get; set; } = "Yes";
@@ -69,6 +85,10 @@ public class ConfirmConfig
     public string CancelText { get; set; } = DefaultUseYesNo ? DefaultNo : DefaultCancelText;
     public string Title { get; set; }
     public string Message { get; set; }
+
+    /// <summary>
+    /// Android only
+    /// </summary>
     public string Icon { get; set; }
     public Action<bool> Action { get; set; }
